pyproject.toml
[tool.poetry]
name = "cointracker"
version = "0.1.0"
description = ""
authors = ["jotan <jonathan.tan27@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
jsonpickle = "^3.0.2"
ngrok = "^0.12.1"
pyngrok = "^7.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
